pdf_platform_deliverables/requirements.txt
flask
requests
python-dotenv
pdfplumber
google-generativeai